+++ lib/executor/common_test.go
@@ -55,7 +55,7 @@ func setupExecutor(t testing.TB, config lib.ExecutorConfig, es *lib.ExecutionSta
 	logEntry := logrus.NewEntry(testLog)
 
 	initVUFunc := func(_ context.Context, logger *logrus.Entry) (lib.InitializedVU, error) {
-		return runner.NewVU(int64(es.GetUniqueVUIdentifier()), engineOut)
+		return runner.NewVU(es.GetUniqueVUIdentifier(), engineOut)
 	}
 	es.SetInitVUFunc(initVUFunc)
